# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StudioFlow - Real Estate Media Delivery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /media/{mediaId}/ai-jobs:
    get:
      operationId: listAiJobs
      tags: [ai-jobs]
      summary: List AI jobs for a media asset
      parameters:
        - in: path
          name: mediaId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiJob"
    post:
      operationId: createAiJob
      tags: [ai-jobs]
      summary: Create AI job (mock)
      parameters:
        - in: path
          name: mediaId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAiJobBody"
      responses:
        "201":
          description: AI job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiJob"
  /ai-jobs/{id}:
    get:
      operationId: getAiJob
      tags: [ai-jobs]
      summary: Get AI job status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiJob"
  /ai-jobs/{id}/cancel:
    patch:
      operationId: cancelAiJob
      tags: [ai-jobs]
      summary: Cancel AI job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiJob"

  # Galleries / Share Links
...
  /media/{mediaId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a media asset
      parameters:
        - in: path
          name: mediaId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create comment
      parameters:
        - in: path
          name: mediaId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Favorites
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # User / Settings
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        businessName:
          type: ["string", "null"]
        plan:
          type: string
          enum: [free, pro, studio]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - plan
        - createdAt
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        businessName:
          type: ["string", "null"]
      required:
        - email
        - password
        - name
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
...
    AiJob:
      type: object
      properties:
        id:
          type: integer
        mediaId:
          type: integer
        jobType:
          type: string
          enum: [virtual_staging, furniture_replacement, declutter, sky_replacement, day_to_dusk, hdr_enhancement, object_removal, color_grading]
        status:
          type: string
          enum: [queued, processing, completed, failed, cancelled]
        progressPercent:
          type: integer
        resultUrl:
          type: ["string", "null"]
        settings:
          type: object
          additionalProperties: true
        errorMessage:
          type: ["string", "null"]
        creditsUsed:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - mediaId
        - jobType
        - status
        - progressPercent
        - creditsUsed
        - createdAt
        - updatedAt
    CreateAiJobBody:
      type: object
      properties:
        jobType:
          type: string
          enum: [virtual_staging, furniture_replacement, declutter, sky_replacement, day_to_dusk, hdr_enhancement, object_removal, color_grading]
        settings:
          type: object
          additionalProperties: true
      required:
        - jobType
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        mediaId:
          type: integer
        authorName:
          type: string
        authorEmail:
          type: ["string", "null"]
        body:
          type: string
        isClientComment:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - mediaId
        - authorName
        - body
        - isClientComment
        - createdAt
    CreateCommentBody:
      type: object
      properties:
        authorName:
          type: string
        authorEmail:
          type: ["string", "null"]
        body:
          type: string
        isClientComment:
          type: boolean
      required:
        - authorName
        - body
...
    Client:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        projectCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - email
        - projectCount
        - createdAt
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - email
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...