# lib/api-zod/src/generated/types/createCommentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudioFlow - Real Estate Media Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCommentBody {
  authorName: string;
  /** @nullable */
  authorEmail?: string | null;
  body: string;
  isClientComment?: boolean;
}