# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudioFlow - Real Estate Media Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from "./userPlan";

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  businessName?: string | null;
  plan: UserPlan;
  createdAt: string;
}